refactor : reorganize package structure under mocka-odm generator module
