+++ mocka-odm/src/main/java/mocka/odm/generator/ODMCreator.java
@@ -1,4 +1,4 @@
-package mocka.odm.generator.odm;
+package mocka.odm.generator;
 
 import mocka.core.GenerateType;
 import mocka.core.VisitedPath;
