+++ mocka-odm/src/test/java/mocka/odm/MockaOdmMainTest.java
@@ -3,8 +3,9 @@ package mocka.odm;
 import mocka.odm.document.Member;
 import mocka.odm.generator.DocumentGenerator;
 import mocka.odm.generator.DocumentGeneratorFactory;
+import mocka.odm.generator.DocumentGeneratorInitializer;
 import mocka.odm.generator.ODMResolver;
-import mocka.odm.generator.odm.ODMCreator;
+import mocka.odm.generator.ODMCreator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
