+++ mocka-odm/src/test/java/mocka/odm/generatorTest/GeoJsonGeneratorTest.java
@@ -1,4 +1,4 @@
-package mocka.odm.generator;
+package mocka.odm.generatorTest;
 
 import mocka.odm.generator.factory.GeoJsonGenerator;
 import mocka.odm.generator.factory.ODMGeneratorFactory;
